Game development and fixes.
Game progress development.Features collisions sound effects animations.

+++ EPL426-main/Assets/Scripts/CollectableRotate.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class CollectableRotate : MonoBehaviour
 {
-    [SerializeField] int rotateSpeed = 1;
+    [SerializeField] float rotateSpeed = 1;
 
     void Update()
     {
